Add files via upload
Here's what I worked on while you guys were gone. We now have the ability to log in an out, sign up, and search items in our search bar. The only task that remains is to filter through the results and find the URL's that do not contain our example words. For example, I'd love for us to be able to have a search bar and a target bar. You search for a product and your target is the company who made them. If the URL does not contain the target, then it will be written in red.